app/requirements.txt
langchain
langchain-core
langchain-community
langchain_experimental

langchain-openai
openai

langchain-anthropic


langchain-google-genai
google-generativeai

langchain-huggingface
transformers
huggingface-hub

python-dotenv

numpy
pandas
scikit-learn
streamlit

# PyPDFLoader
youtube-transcript-api
duckduckgo-search


fastapi 
uvicorn
sentence-transformers 
faiss-cpu
pypdf
python-multipart
docx2txt